requirements.txt
streamlit
pandas
pillow
pytesseract
pdf2image
pypdf